@@ -1,10 +1,5 @@
 package com.koli.openquiz.model;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
@@ -12,69 +7,19 @@ public class Dictionary {
 
     private String name;
     private int version;
-    private List<Word> content;
-
-    public static Dictionary createDictionary(final JSONObject jsonObject) {
-        try {
-            final Dictionary dictionary = new Dictionary();
-            dictionary.setName(jsonObject.getString("name"));
-            dictionary.setVersion(jsonObject.getInt("version"));
-            dictionary.setContent(parseContent(new JSONArray(jsonObject.getString("dictionary"))));
-            return dictionary;
-        } catch (JSONException e) {
-            throw new RuntimeException();
-        }
-    }
-
-    public static Dictionary createDictionary(final JSONArray jsonArray) throws JSONException {
-        final Dictionary dictionary = new Dictionary();
-        dictionary.setName("Unknown");
-        dictionary.setVersion(1);
-        dictionary.setContent(parseContent(jsonArray));
-        return dictionary;
-    }
-
-    private static List<Word> parseContent(final JSONArray jsonArray) throws JSONException {
-        final List<Word> content = new ArrayList<>();
-
-        for (int i = 0; i < jsonArray.length(); i++) {
-            final JSONObject jsonWordObject = jsonArray.getJSONObject(i);
-            content.add(Word.createWord(jsonWordObject));
-        }
-        return content;
-    }
-
-    public JSONObject toJson() {
-        JSONObject jsonObject = new JSONObject();
-        try {
-            jsonObject.put("name", name);
-            jsonObject.put("version", version);
-            jsonObject.put("dictionary", convertContent());
-        } catch (JSONException e) {
-            jsonObject = null;
-        }
-        return jsonObject;
-    }
-
-    private JSONArray convertContent() {
-        JSONArray jsonArray = new JSONArray();
-        for(Word word : content) {
-            jsonArray.put(word.toJson());
-        }
-        return jsonArray;
-    }
+    private List<Word> dictionary;
 
     public Word getRandomWord() {
         final Random random = new Random();
-        return content.get(random.nextInt(content.size()));
+        return dictionary.get(random.nextInt(dictionary.size()));
     }
 
     public int size() {
-        return content.size();
+        return dictionary.size();
     }
 
     public Word get(final int index) {
-        return content.get(index);
+        return dictionary.get(index);
     }
 
     public String getName() {
@@ -93,12 +38,20 @@ public class Dictionary {
         this.version = version;
     }
 
-    public List<Word> getContent() {
-        return content;
+    public List<Word> getDictionary() {
+        return dictionary;
     }
 
-    public void setContent(final List<Word> content) {
-        this.content = content;
+    public void setDictionary(final List<Word> dictionary) {
+        this.dictionary = dictionary;
     }
 
+    @Override
+    public String toString() {
+        return "Dictionary{" +
+            "name='" + name + '\'' +
+            ", version=" + version +
+            ", dictionary=" + dictionary +
+            '}';
+    }
 }
