@@ -4,34 +4,42 @@ import android.content.Context;
 
 import com.koli.openquiz.model.Dictionary;
 import com.koli.openquiz.model.Question;
-import com.koli.openquiz.model.QuizSettings;
+import com.koli.openquiz.settings.QuizSettings;
 import com.koli.openquiz.model.Word;
 import com.koli.openquiz.model.exception.InvalidDictionaryException;
+import com.koli.openquiz.persistence.DictionaryStore;
+import com.koli.openquiz.settings.SettingsProvider;
+import com.squareup.moshi.JsonAdapter;
+import com.squareup.moshi.Moshi;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
+import java.util.logging.Logger;
 
 public class QuestionProvider {
 
-    private Dictionary dictionary;
-    private QuizSettings settings;
-    private Random random;
+    private final Logger log = Logger.getLogger(QuestionProvider.class.getName());
+
+    private final Dictionary dictionary;
+    private final SettingsProvider settings;
+    private final Random random;
 
     public QuestionProvider(Context context, String filename) {
         this.random = new Random();
-        this.settings = new QuizSettings(context);
+        this.settings = new SettingsProvider(context);
+        Moshi moshi = new Moshi.Builder().build();
+        JsonAdapter<Dictionary> adapter = moshi.adapter(Dictionary.class);
 
-        final DictionaryStorageService service = new DictionaryStorageService(context);
+        final DictionaryStore service = new DictionaryStore(context);
+
         try {
-            final JSONObject jsonObject = new JSONObject(service.read(filename));
-            this.dictionary = Dictionary.createDictionary(jsonObject);
-        } catch (JSONException e) {
-            throw new InvalidDictionaryException();
+            String fileContent = service.read(filename);
+            this.dictionary = adapter.lenient().fromJson(fileContent);
+            log.info(String.format("Initialized with dictionary: %s", dictionary));
+        } catch (IOException e) {
+            throw new InvalidDictionaryException(e);
         }
     }
 
@@ -53,7 +61,7 @@ public class QuestionProvider {
     }
 
     private int getPossibleChoiceCount() {
-        return (dictionary.size() < settings.getChoiceCount()) ? dictionary.size() : settings.getChoiceCount();
+        return Math.min(dictionary.size(), settings.readInt(QuizSettings.CHOICE_COUNT));
     }
 
     private Word getRandomWord(List<Word> choices) {
