@@ -1,58 +1,60 @@
 package com.koli.openquiz.view.settings;
 
-import android.app.Activity;
 import android.app.Dialog;
-import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.EditText;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
+import androidx.fragment.app.DialogFragment;
+import androidx.fragment.app.FragmentManager;
 
 import com.koli.openquiz.R;
 
 public class AddDictionaryDialogFragment extends DialogFragment {
 
     public interface Listener {
-        void onSaveFile(String name);
+        void onNameEntered(String name);
     }
 
-    private Activity activity;
+    private final Listener listener;
     private View view;
-    private Listener listener;
 
+    public AddDictionaryDialogFragment(Listener listener) {
+        this.listener = listener;
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        super.onCreateView(inflater, container, savedInstanceState);
+        this.view = inflater.inflate(R.layout.fragment_dialog_add_dictionary, container, false);
+        return view;
+    }
+
+    @NonNull
     @Override
     public Dialog onCreateDialog(Bundle savedInstance) {
+        super.onCreateDialog(savedInstance);
 
-        activity = getActivity();
-        LayoutInflater inflater = activity.getLayoutInflater();
-        view = inflater.inflate(R.layout.fragment_dialog_add_dictionary, null);
-
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-
-        builder.setView(view)
-                .setPositiveButton("Save", (dialog, which) -> {
-                    EditText editText = (EditText) view.findViewById(R.id.dictionary_add_name);
-                    listener.onSaveFile(editText.getText().toString());
-                    activity.getFragmentManager().popBackStack();
-                });
-
-        return builder.create();
+        return new AlertDialog.Builder(requireActivity())
+            .setView(view)
+            .setPositiveButton("Save", (dialog, which) -> {
+                EditText editText = view.findViewById(R.id.dictionary_add_name);
+                listener.onNameEntered(editText.getText().toString());
+                getParentFragmentManager().popBackStack();
+            })
+            .create();
     }
 
     @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        listener = (Listener) context;
-    }
-
-    @Override
-    public void show (FragmentManager manager, String tag) {
-        FragmentTransaction ft = manager.beginTransaction();
-        ft.add(this, tag);
-        ft.commitAllowingStateLoss();
+    public void show(FragmentManager manager, String tag) {
+        manager.beginTransaction()
+            .add(this, tag)
+            .commitAllowingStateLoss();
     }
 }
